partitionedconsumergroups: clean up dependencies
