+++ .github/workflows/partitionedconsumergroups.yaml
@@ -35,7 +35,9 @@ jobs:
         run: |
           $(exit $(go fmt  ./... | wc -l))
           go vet ./...
+          go vet ./test/...
           staticcheck ./...
+          staticcheck ./test/...
           find . -type f -name "*.go" | xargs misspell -error -locale US
 
   test:
@@ -54,3 +56,4 @@ jobs:
         shell: bash --noprofile --norc -x -eo pipefail {0}
         run: |
           go test -v -count=1 ./...
+          go test -v -count=1 ./test/...
