+++ partitionedconsumergroups/go.work
@@ -1,0 +1,6 @@
+go 1.23.7
+
+use (
+	.
+	./test
+)
