+++ partitionedconsumergroups/stream_consumer_group.go
@@ -11,15 +11,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package streamconsumergroup
+package partitionedconsumergroups
 
 import (
 	"context"
 	"fmt"
-	"github.com/nats-io/nats.go"
-	"github.com/nats-io/nats.go/jetstream"
 	"slices"
 	"time"
+
+	"github.com/nats-io/nats.go"
+	"github.com/nats-io/nats.go/jetstream"
 )
 
 // Contains the things that are common to both types of consumer groups
